+++ src/Server/Protocol/deserializer.cpp
@@ -97,7 +97,7 @@ std::unique_ptr<CreateGameDTO> Deserializer::deserializeCreateGame(bool& wasClos
 }
 
 std::unique_ptr<MapsListDTO> Deserializer::deserializeMapsList(bool& wasClosed,
-                                                               uint32_t& playerId) {
+                                                               uint8_t& playerId) {
     std::cout << "[SERVER DESERIALIZER] Deserialize Maps List" << std::endl;
     return std::make_unique<MapsListDTO>();
 }
@@ -125,7 +125,7 @@ std::unique_ptr<CommandDTO> Deserializer::deserializeGamesList(bool& wasClosed, 
     return std::make_unique<GamesListDTO>();
 }
 
-std::unique_ptr<GameCommandDTO> Deserializer::deserializeMove(bool& wasClosed, uint32_t& playerId) {
+std::unique_ptr<GameCommandDTO> Deserializer::deserializeMove(bool& wasClosed, uint8_t& playerId) {
     Direction direction;
     socket->recvall(&direction, sizeof(char), &wasClosed);
     std::cout << "[SERVER DESERIALIZER MOVE] Received direction: " << (int)direction << std::endl;
